Image recognition script
This uses  the following libraries:
Tensorflow
Keras
Pillow
Matplotlib
Numpy

Using a large data model and deep learning, this script allows the data model to be trained to identify images from the CIFAR-10 dataset. 


Please note this script will not run properly if you download and attempt to run it. Please ensure you've downloaded to proper libraries and that you have set your libraries in the correct interpreter folder too, otherwise this will not work. Lastly, this is adaptable code, there are lines in which you need to replace it with your dependant variable to train the model accordingly. You may also need to expand the function "def prediction(self, image_path): if you wish to compare multiple images at once.